fix footer , other padding on landing , fix signup and singin modals , change landing page route and add recaptcha
