+++ src/components/auth_model/index.js
@@ -1,6 +1,6 @@
-import Modal from '@mui/material/Modal';
-import Box from '@mui/material/Box';
-import './styles.css'
+import Modal from "@mui/material/Modal";
+import Box from "@mui/material/Box";
+import "./styles.css";
 
 const AuthModel = ({ open, children, handleModel, details }) => {
   return (
@@ -8,25 +8,27 @@ const AuthModel = ({ open, children, handleModel, details }) => {
       open={open}
       onClose={handleModel}
       sx={{
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center'
-    }}>
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
       <Box
-        className='auth-model'
+        className="auth-model"
         sx={{
-          backgroundColor: '#fff',
-          maxWidth: '430px'
-        }}>
+          backgroundColor: "#fff",
+          maxWidth: "430px",
+        }}
+      >
         <div className="auth-heading">
-          <h3>ComposeTrip</h3>
+          <h4>ComposeTrip</h4>
         </div>
-        <div className='auth-box'>
-            <div className="auth-title">
-                <h6>{details.title}</h6>
-                <p>{details.text}</p>
-            </div>
-            {children}
+        <div className="auth-box">
+          <div className="auth-title">
+            <h6>{details.title}</h6>
+            <p>{details.text}</p>
+          </div>
+          {children}
         </div>
       </Box>
     </Modal>
