+++ src/components/auth_model/styles.css
@@ -1,61 +1,77 @@
 .auth-model .auth-heading {
-    text-transform: uppercase;
-    color: #2B92D5;
-    font-size: 30px;
-    font-weight: 800;
-    line-height: 43px;
-    letter-spacing: 0em;
-    padding: 20px 40px;
-    border-bottom: 1px solid #D2D4DA;
+  text-transform: uppercase;
+  color: #2b92d5;
+  font-size: 30px;
+  font-weight: 800;
+  line-height: 43px;
+  letter-spacing: 0em;
+  padding: 20px;
+  border-bottom: 1px solid #d2d4da;
 }
 
 .auth-model .auth-box {
-    padding: 20px 40px 50px;
+  padding: 20px 20px 50px;
 }
 
 .auth-model .auth-title h6 {
-    font-size: 26px;
-    font-weight: 700;
-    line-height: 43px;
-    letter-spacing: 0em;
+  font-size: 26px;
+  font-weight: 700;
+  line-height: 43px;
+  letter-spacing: 0em;
 }
 
 .auth-model .auth-title h6 span {
-    color: #2B92D5;
+  color: #2b92d5;
 }
 
 .auth-model .auth-title p {
-    font-size: 16px;
-    font-weight: 600;
+  font-size: 16px;
+  font-weight: 600;
 }
 
 /* Auth Pages Modifiying CSS */
 .auth-model label {
-    font-size: 16px;
-    font-weight: 600;
-    line-height: 27px;
-    letter-spacing: 0em;
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 27px;
+  letter-spacing: 0em;
 }
 
-.auth-model .css-8j6b76-MuiInputBase-root-MuiOutlinedInput-root, .auth-model .css-md26zr-MuiInputBase-root-MuiOutlinedInput-root {
-    border-radius: 100px;
+.auth-model .css-8j6b76-MuiInputBase-root-MuiOutlinedInput-root,
+.auth-model .css-md26zr-MuiInputBase-root-MuiOutlinedInput-root {
+  border-radius: 100px;
 }
 
 .auth-model .css-8j6b76-MuiInputBase-root-MuiOutlinedInput-root svg {
-    color: #2B92D5;
+  color: #2b92d5;
 }
 
 .auth-model .css-8j6b76-MuiInputBase-root-MuiOutlinedInput-root input {
-    background-color: #fff !important;
-    padding: 15px 20px !important;
+  background-color: #fff !important;
+  padding: 15px 20px !important;
 }
 
 .auth-model button {
-    background: #2B92D5;
-    border-radius: 100px;
-    padding: 15px 20px;
+  background: #2b92d5;
+  border-radius: 100px;
+  padding: 15px 20px;
 }
 
 .auth-model .google-button > div {
-    margin-top: 0px !important;
+  margin-top: 0px !important;
+}
+
+@media screen and (max-width: 992px) {
+  .auth-model {
+    max-height: 600px;
+    overflow: auto;
+    height: 100%;
+  }
+}
+@media screen and (max-width: 560px) {
+  .auth-model {
+    max-height: 100%;
+    overflow: auto;
+    height: 100%;
+  }
 }
