+++ src/components/thirdBlock.js
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import Box from "@mui/material/Box";
 import AppBar from "@/components/navbar";
 import { Typography } from "@mui/material";
-import Icon from "@mui/material";
+import { Container } from "@mui/material";
 import { useRouter } from "next/navigation";
 import * as COLORS from "@/constants/colors";
 import blockImage from "@/assets/images/thirdBlock.svg";
@@ -17,83 +17,88 @@ function ThirdBlock() {
   return (
     <Box
       sx={{
-        display: "flex",
-        flexDirection: {
-          xs: "column",
-          md: "row",
-          lg: "row",
-        },
-        justifyContent: "center",
-        alignItems: "center",
         width: "100%",
-        padding: {
-          xs: "30px 30px",
-          md: "30px 0px 30px 40px",
-          lg: "30px 0px 30px 100px",
-        },
+        paddingTop: "50px",
+        paddingBottom: "50px",
         backgroundColor: COLORS.white,
       }}
     >
-      <Box
-        sx={{
-          flexDirection: {
-            xs: "column",
-            md: "row",
-            lg: "row",
-          },
-        }}
-      >
-        <Typography
+      <Container>
+        <Box
           sx={{
-            fontSize: { xs: "22px", md: "32px", lg: "32px" },
-            fontWeight: "800",
-            fontFamily: "raleway",
-            textTransform: "uppercase",
-            color: COLORS.black,
-            textAlign: { md: "center", xs: "center", lg: "left" },
+            display: "flex",
+            flexDirection: {
+              xs: "column",
+              md: "row",
+              lg: "row",
+            },
+            justifyContent: "center",
+            alignItems: "center",
           }}
         >
-          <Typography
-            as="span"
+          <Box
             sx={{
-              fontSize: { xs: "22px", md: "32px", lg: "32px" },
-              fontWeight: "700",
-              fontFamily: "raleway",
-              textTransform: "uppercase",
-              color: COLORS.primary,
+              flexDirection: {
+                xs: "column",
+                md: "row",
+                lg: "row",
+              },
             }}
           >
-            AI-Curated&nbsp;
-          </Typography>
-          Itineraries
-        </Typography>
-        <Typography
-          sx={{
-            fontSize: { xs: "16px", md: "18px", lg: "22px" },
-            fontWeight: "500",
-            fontFamily: "raleway",
-            color: COLORS.textColor,
-            textAlign: { md: "center", xs: "center", lg: "left" },
-            width: { xs: "100%", md: "100%", lg: "75%" },
-          }}
-        >
-          Are you an urban explorer, a nature enthusiast, or a fervent history
-          aficionado? Share with us your travel inclinations and we'll identify
-          the best destinations that align with your preferences. With
-          ComposeTrip, you can design trips that truly resonate with you.
-        </Typography>
-      </Box>
-      <Box
-        sx={{
-          flex: 1,
-          display: "flex",
-          flexDirection: "row",
-          justifyContent: "flex-end",
-          marginTop: { xs: "20px", md: "0px", lg: "0px" },
-        }}
-      >
-        <Image src={blockImage} alt="Image" />
-      </Box>
+            <Typography
+              sx={{
+                fontSize: { xs: "22px", md: "32px", lg: "32px" },
+                fontWeight: "800",
+                fontFamily: "raleway",
+                textTransform: "uppercase",
+                color: COLORS.black,
+                textAlign: { md: "center", xs: "center", lg: "left" },
+              }}
+            >
+              <Typography
+                as="span"
+                sx={{
+                  fontSize: { xs: "22px", md: "32px", lg: "32px" },
+                  fontWeight: "700",
+                  fontFamily: "raleway",
+                  textTransform: "uppercase",
+                  color: COLORS.primary,
+                }}
+              >
+                AI-Curated&nbsp;
+              </Typography>
+              Itineraries
+            </Typography>
+            <Typography
+              sx={{
+                fontSize: { xs: "16px", md: "18px", lg: "22px" },
+                fontWeight: "500",
+                fontFamily: "raleway",
+                color: COLORS.textColor,
+                textAlign: { md: "center", xs: "center", lg: "left" },
+                width: { xs: "100%", md: "100%", lg: "75%" },
+              }}
+            >
+              Are you an urban explorer, a nature enthusiast, or a fervent
+              history aficionado? Share with us your travel inclinations and
+              we'll identify the best destinations that align with your
+              preferences. With ComposeTrip, you can design trips that truly
+              resonate with you.
+            </Typography>
+          </Box>
+          <Box
+            sx={{
+              flex: 1,
+              display: "flex",
+              flexDirection: "row",
+              justifyContent: "flex-end",
+              marginTop: { xs: "20px", md: "0px", lg: "0px" },
+            }}
+          >
+            <Image src={blockImage} alt="Image" />
+          </Box>
+        </Box>
+      </Container>
     </Box>
   );
 }
